factor shared management config into separate file
