+++ google/resource_container_cluster.go
@@ -261,27 +261,7 @@ func resourceContainerCluster() *schema.Resource {
 							Computed: true,
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
-									"management": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Computed: true,
-										MaxItems: 1,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"auto_repair": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Default:  false,
-												},
-
-												"auto_upgrade": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Default:  false,
-												},
-											},
-										},
-									},
+									"management": schemaManagement,
 									"oauth_scopes": {
 										Type:             schema.TypeList,
 										Optional:         true,
@@ -1916,28 +1896,10 @@ func expandAutoProvisioningDefaults(configured interface{}, d *schema.ResourceDa
 	config := l[0].(map[string]interface{})
 
 	return &containerBeta.AutoprovisioningNodePoolDefaults{
-		Management:     expandManagement(config["management"], d),
+		Management:     expandManagement(config["management"]),
 		OauthScopes:    convertStringArr(config["oauth_scopes"].([]interface{})),
 		ServiceAccount: config["service_account"].(string),
 	}
-}
-
-func expandManagement(configured interface{}, d *schema.ResourceData) *containerBeta.NodeManagement {
-	l, ok := configured.([]interface{})
-	if !ok || l == nil || len(l) == 0 || l[0] == nil {
-		return &containerBeta.NodeManagement{}
-	}
-	config := l[0].(map[string]interface{})
-	result := &containerBeta.NodeManagement{
-		ForceSendFields: []string{"AutoRepair", "AutoUpgrade"},
-	}
-	if v, ok := config["auto_repair"]; ok {
-		result.AutoRepair = v.(bool)
-	}
-	if v, ok := config["auto_upgrade"]; ok {
-		result.AutoUpgrade = v.(bool)
-	}
-	return result
 }
 
 func expandAuthenticatorGroupsConfig(configured interface{}) *containerBeta.AuthenticatorGroupsConfig {
@@ -2252,14 +2214,6 @@ func flattenAutoProvisioningDefaults(a *containerBeta.AutoprovisioningNodePoolDe
 	r["management"] = flattenManagement(a.Management)
 	r["oauth_scopes"] = a.OauthScopes
 	r["service_account"] = a.ServiceAccount
-
-	return []map[string]interface{}{r}
-}
-
-func flattenManagement(n *containerBeta.NodeManagement) []map[string]interface{} {
-	r := make(map[string]interface{})
-	r["auto_repair"] = n.AutoRepair
-	r["auto_upgrade"] = n.AutoUpgrade
 
 	return []map[string]interface{}{r}
 }
